introduce use of mapAsync
