+++ packages/core/utils/lib/async.js
@@ -20,7 +20,33 @@ function pipeAsync(...methods) {
  */
 const mapAsync = curry(pMap);
 
+/**
+ * Because of how mysql works, making parallel requests can cause deadlocks.
+ * This function will run the requests sequentially if mysql is used. Else,
+ * it will encapsulate them in a Promise.all.
+ *
+ * @type { import('./async').MapAsync }
+ */
+const mapAsyncDialects = async (array, func) => {
+  const results = [];
+
+  switch (strapi.db.dialect.client) {
+    case 'mysql': {
+      mapAsync(array, async (elm) => {
+        results.push(await func(elm));
+      });
+      break;
+    }
+    default:
+      results.push(...(await Promise.all(array.map(func))));
+      break;
+  }
+
+  return results;
+};
+
 module.exports = {
   mapAsync,
+  mapAsyncDialects,
   pipeAsync,
 };
